+++ python/src/opendp/extras/numpy/_make_np_count/__init__.py
@@ -1,11 +1,12 @@
-from opendp.mod import Domain, Metric, Measurement
+from opendp.mod import Domain, Metric, Transformation
 import opendp.prelude as dp
 from opendp.extras._utilities import to_then, with_privacy
+from opendp._internal import _make_transformation
 
 # planning to make this public, but may make more API changes
 
 
-def make_np_count(input_domain: Domain, input_metric: Metric) -> Measurement:
+def make_np_count(input_domain: Domain, input_metric: Metric) -> Transformation:
     """Construct a Transformation that returns the length of axis 0 of a 2-dimensional array.
 
     :param input_domain: instance of `array2_domain(...)`
@@ -17,7 +18,7 @@ def make_np_count(input_domain: Domain, input_metric: Metric) -> Measurement:
     if input_metric != dp.symmetric_distance():
         raise ValueError("input metric must be symmetric distance")
 
-    return dp.t._make_transformation(
+    return _make_transformation(
         input_domain,
         input_metric,
         dp.atom_domain(T=int),
