+++ python/src/opendp/extras/sklearn/_make_eigenvector/__init__.py
@@ -2,6 +2,7 @@ from opendp.extras.numpy import _sscp_domain
 from opendp.extras._utilities import to_then
 from opendp._lib import get_np_csprng, import_optional_dependency
 from opendp.mod import Domain, Metric, Transformation, Measurement
+from opendp._internal import _make_measurement, _make_transformation
 
 # planning to make this public, but may make more API changes
 
@@ -79,7 +80,7 @@ def make_private_eigenvector(
             if np.exp(-u.T @ A @ u) / (M * (u.T @ Omega @ u) ** (d / 2)):
                 return u
 
-    return dp.m._make_measurement(
+    return _make_measurement(
         input_domain,
         input_metric,
         dp.max_divergence(),
@@ -118,7 +119,7 @@ def make_np_sscp_projection(
         )
 
     kwargs = input_desc._asdict() | {"num_features": P.shape[0]}
-    return dp.t._make_transformation(
+    return _make_transformation(
         input_domain,
         input_metric,
         _sscp_domain(**kwargs),
@@ -194,7 +195,7 @@ def make_private_eigenvectors(
 
         return theta.T
 
-    return dp.m._make_measurement(
+    return _make_measurement(
         input_domain,
         input_metric,
         m_compose.output_measure,
