+++ rust/src/transformations/make_user_transformation/mod.rs
@@ -66,70 +66,3 @@ pub extern "C" fn opendp_transformations__make_user_transformation(
     )
     .into()
 }
-
-#[bootstrap(
-    name = "_make_transformation",
-    arguments(
-        input_domain(hint = "Domain"),
-        input_metric(hint = "Metric"),
-        output_domain(hint = "Domain"),
-        output_metric(hint = "Metric"),
-        function(rust_type = "$domain_carrier_type(output_domain)"),
-        stability_map(rust_type = "$metric_distance_type(output_metric)"),
-    ),
-    dependencies(
-        "input_domain",
-        "input_metric",
-        "output_domain",
-        "output_metric",
-        "c_function",
-        "c_stability_map"
-    )
-)]
-/// Construct a Transformation from user-defined callbacks.
-/// This is meant for internal use, as it does not require "honest-but-curious",
-/// unlike `make_user_transformation`.
-///
-/// This is used when you write a constructor in a bindings language that does not have preconditions,
-/// and is meant to be a part of the public API for the OpenDP Library.
-///
-/// For every element $x$ in `input_domain`,
-/// `function(x)` must be a member of `output_domain` or raise a data-independent runtime exception.
-///
-/// `function` must not have side-effects and only raise data-independent exceptions,
-/// and `stability_map` must be a pure function.
-///
-/// # Arguments
-/// * `input_domain` - A domain describing the set of valid inputs for the function.
-/// * `input_metric` - The metric from which distances between adjacent inputs are measured.
-/// * `output_domain` - A domain describing the set of valid outputs of the function.
-/// * `output_metric` - The metric from which distances between outputs of adjacent inputs are measured.
-/// * `function` - A function mapping data from `input_domain` to `output_domain`.
-/// * `stability_map` - A function mapping distances from `input_metric` to `output_metric`.
-///
-/// # Proof Definition
-/// Returns a valid transformation where for every pair of elements $x, x'$ in `input_domain`,
-/// and for every pair `(d_in, d_out)`,
-/// where `d_in` has the associated type for `input_metric` and `d_out` has the associated type for `output_metric`,
-/// if $x, x'$ are `d_in`-close under `input_metric`, `stability_map(d_in)` does not raise an exception,
-/// and `stability_map(d_in) <= d_out`,
-/// then `function(x), function(x')` are d_out-close under `output_metric`.
-#[no_mangle]
-pub extern "C" fn opendp_transformations___make_transformation(
-    input_domain: *const AnyDomain,
-    input_metric: *const AnyMetric,
-    output_domain: *const AnyDomain,
-    output_metric: *const AnyMetric,
-    function: CallbackFn,
-    stability_map: CallbackFn,
-) -> FfiResult<*mut AnyTransformation> {
-    Transformation::new(
-        try_as_ref!(input_domain).clone(),
-        try_as_ref!(output_domain).clone(),
-        Function::new_fallible(wrap_func(function)),
-        try_as_ref!(input_metric).clone(),
-        try_as_ref!(output_metric).clone(),
-        StabilityMap::new_fallible(wrap_func(stability_map)),
-    )
-    .into()
-}
